chore(bin): Init Subcommand and Refactor
