+++ bin/mev/src/cmd.rs
@@ -1,7 +1,9 @@
+pub mod debug;
+pub mod init;
+
 #[cfg(feature = "boost")]
 pub mod boost;
 #[cfg(feature = "build")]
 pub mod build;
-pub mod config;
 #[cfg(feature = "relay")]
 pub mod relay;
