+++ bin/mev/src/config.rs
@@ -1,18 +1,17 @@
-use clap::Args;
 use ethereum_consensus::networks::Network;
 use eyre::WrapErr;
+use mev_rs::config::from_toml_file;
+use serde::{Deserialize, Serialize};
+use std::{fmt, path::Path};
+
 #[cfg(feature = "boost")]
 use mev_boost_rs::Config as BoostConfig;
 #[cfg(feature = "build")]
 use mev_build_rs::reth_builder::Config as BuildConfig;
 #[cfg(feature = "relay")]
 use mev_relay_rs::Config as RelayConfig;
-use mev_rs::config::from_toml_file;
-use serde::Deserialize;
-use std::{fmt, path::Path};
-use tracing::info;
 
-#[derive(Debug, Deserialize)]
+#[derive(Debug, Serialize, Deserialize)]
 pub struct Config {
     pub network: Network,
     #[cfg(feature = "boost")]
@@ -29,23 +28,5 @@ impl Config {
         tracing::info!("loading config from `{path}`...");
 
         from_toml_file::<_, Self>(path.as_ref()).wrap_err("could not parse TOML")
-    }
-}
-
-#[derive(Debug, Args)]
-#[clap(about = "🔬 (debug) utility to verify configuration")]
-pub struct Command {
-    #[clap(env)]
-    config_file: String,
-}
-
-impl Command {
-    pub async fn execute(self) -> eyre::Result<()> {
-        let config_file = self.config_file;
-
-        let config = Config::from_toml_file(config_file)?;
-        info!("{config:#?}");
-
-        Ok(())
     }
 }
