+++ bin/mev/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod cmd;
+pub mod config;
